requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
langchain-core
openai
duckduckgo-search
asknews
tiktoken
requests
re
